do not update muesli.yml if it exists during docker compose startup

+++ docker-serve.sh
@@ -1,5 +1,5 @@
 #!/usr/bin/env bash
-if [[ -v MUESLI_TESTMODE ]]
+if [[ -v MUESLI_TESTMODE ]] && [[ ! -f muesli.yml ]]
 then
     sed 's/\/\/\//\/\/postgres@postgres\//' muesli.yml.example | sed 's/production: True/production: False/' > muesli.yml
     sed 's/\/\/\//\/\/postgres@postgres\//' alembic.ini.example > alembic.ini
